Add a way to connect to an unsupported target
Though JLink supports many SoC, it doesn't support all of them,
especially when there new or in preproduction.
This adds a way to load JLink scripts, usually used to configure
JLink to work with unsupported target.
In addition, this updates JLink.connect() to be able to connect
without a target name.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>